archive/requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.3
db-dtypes==1.1.1
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.3.2
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nbformat==5.9.2
nest-asyncio==1.5.6
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
pip==23.2.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.15.0
plotly-express==0.4.1
pooch==1.7.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
rsa==4.9
scipy==1.11.1
setuptools==68.0.0
six==1.16.0
stack-data==0.6.2
statsmodels==0.14.0
tenacity==8.2.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
wheel==0.41.0
zipp==3.16.2
